refactored prd agent
